requirements.txt
fastapi[standard]
python-escpos[all]